+++ .github/workflows/tests.yml
@@ -36,7 +36,7 @@ jobs:
       run: uv sync --group dev
 
     - name: Run tests
-      run: uv run pytest --cov=blueprints --cov-report=xml
+      run: uv run pytest --cov=datarecipes --cov-report=xml
 
     - name: Upload coverage
       env:
