--- blueprints/__init__.py
@@ -1,24 +1,0 @@
-# -*- coding: utf-8 -*-
-
-"""Top-level package for blueprints."""
-
-__author__ = """Tom Rutherford"""
-__email__ = "foreverwintr@gmail.com"
-__version__ = "0.1.0"
-
-
-from blueprints.blueprint import Blueprint
-from blueprints.factory import Factory
-from blueprints.factory import FactoryMP
-from blueprints.recipes.base import Dependencies
-from blueprints.recipes.base import DependencyRequest
-from blueprints.recipes.base import Recipe
-
-__all__ = [
-    "Blueprint",
-    "DependencyRequest",
-    "Dependencies",
-    "Recipe",
-    "Factory",
-    "FactoryMP",
-]
