--- blueprints/recipes/__init__.py
@@ -1,13 +1,0 @@
-from blueprints.recipes.general import FromFunction
-from blueprints.recipes.general import Object
-from blueprints.recipes.static_frame import FrameFromDelimited
-from blueprints.recipes.static_frame import FrameFromRecipes
-from blueprints.recipes.static_frame import SeriesFromDelimited
-
-__all__ = [
-    "Object",
-    "FromFunction",
-    "SeriesFromDelimited",
-    "FrameFromDelimited",
-    "FrameFromRecipes",
-]
