+++ datarecipes/exceptions.py
@@ -1,10 +1,10 @@
-class blueprintsError(Exception):
+class datarecipesError(Exception):
     """Base class for frame factory errors"""
 
 
-class ConfigurationError(blueprintsError):
+class ConfigurationError(datarecipesError):
     """The specified configuration is invalid"""
 
 
-class MissingDependencyError(blueprintsError):
+class MissingDependencyError(datarecipesError):
     """An upstream dependency is missing and at least one dependency does not allow missing"""
