+++ datarecipes/renderers/dash_renderer.py
