+++ datarecipes/serialization.py
@@ -6,9 +6,9 @@ import typing as tp
 import networkx as nx
 from frozendict import frozendict
 
-from blueprints import util
-from blueprints.recipes.base import RECIPE_TYPE_REGISTRY
-from blueprints.recipes.base import Recipe
+from datarecipes import util
+from datarecipes.recipes.base import RECIPE_TYPE_REGISTRY
+from datarecipes.recipes.base import Recipe
 
 
 class RecipeRegistry:
