+++ tox.ini
@@ -6,4 +6,4 @@ isolated_build = true
 allowlist_externals = uv
 commands =
     uv sync --group dev
-    uv run pytest --cov=blueprints --cov-report=xml
+    uv run pytest --cov=datarecipes --cov-report=xml
